pyproject.toml
[project]
name = "bettery-validator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cosmospy>=6.0.0",
    "black>=26.1.0",
    "isort>=7.0.0",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.19",
    "fastapi>=0.129.0",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.71.2",
    "uvicorn>=0.41.0",
    "protobuf>=6.33.5",
    "asyncpg>=0.31.0",
]
